created Login page, exported Logins component and rendered it on this page
