+++ client/src/App.js
@@ -10,7 +10,7 @@ import "./App.css"
 // Pages
 import Home from "./pages/Home";
 import NoMatch from "./pages/NoMatch";
-import Logins from "./components/LoginSign/Logins";
+import Login from "./pages/Login";
 
 function App() {
   return (
@@ -20,7 +20,7 @@ function App() {
         <Switch>
           <Route exact path="/" component={Home} />
           <Route exact path="/home" component={Home} />
-          <Route exact path="/logins" component={Logins} />
+          <Route exact path="/login" component={Login} />
           <Route component={NoMatch} />
         </Switch> 
         <Footer />
