+++ client/src/pages/Login.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import Logins from "../components/LoginSign/Logins"
+
+export default function Login() {
+    return (
+        <div>
+            <Logins />
+        </div>
+    )
+}
